Rename transcribe to generateTranscription.
