+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changes
 
+- **breaking change**: renamed `transcribe` to `generateTranscription`
 - **breaking change**: renamed `synthesizeSpeech` to `generateSpeech`
 - **breaking change**: renamed `LmntSpeechSynthesisModel` to `LmntSpeechModel`
 - **breaking change**: renamed `ElevenLabesSpeechSynthesisModel` to `ElevenLabsSpeechModel`
