+++ examples/basic/src/model-function/generate-transcription-example.ts
@@ -1,5 +1,5 @@
-import { OpenAITranscriptionModel, transcribe } from "modelfusion";
 import dotenv from "dotenv";
+import { OpenAITranscriptionModel, generateTranscription } from "modelfusion";
 import fs from "node:fs";
 
 dotenv.config();
@@ -7,12 +7,9 @@ dotenv.config();
 async function main() {
   const data = await fs.promises.readFile("data/test.mp3");
 
-  const transcription = await transcribe(
+  const transcription = await generateTranscription(
     new OpenAITranscriptionModel({ model: "whisper-1" }),
-    {
-      type: "mp3",
-      data,
-    }
+    { type: "mp3", data }
   );
 
   console.log(transcription);
