+++ examples/basic/src/model-provider/openai/openai-generate-text-api-configuration-example.ts
