+++ examples/basic/src/util/cost/embed-cost-calculation-example.ts
