+++ src/model-function/generate-transcription/TranscriptionModel.ts
