+++ src/model-function/index.ts
@@ -27,8 +27,8 @@ export * from "./generate-structure/generateStructure.js";
 export * from "./generate-structure/generateStructureOrText.js";
 export * from "./generate-structure/streamStructure.js";
 export * from "./generate-text/index.js";
+export * from "./generate-transcription/TranscriptionEvent.js";
+export * from "./generate-transcription/TranscriptionModel.js";
+export * from "./generate-transcription/generateTranscription.js";
 export * from "./tokenize-text/Tokenizer.js";
 export * from "./tokenize-text/countTokens.js";
-export * from "./transcribe-speech/TranscriptionEvent.js";
-export * from "./transcribe-speech/TranscriptionModel.js";
-export * from "./transcribe-speech/transcribe.js";
